+++ FBNDP-MLTN/Data/DataCenter.cpp
@@ -12,8 +12,8 @@ void DataCenter::SetNodeData(const FileDataList& InData)
 		}
 
 		NodeData Data;
-		Data.Num = static_cast<uint32_t>(stoul(RowData.at(1)));
-		Data.Coord = Coordinate(static_cast<uint32_t>(stoul(RowData.at(2))), static_cast<uint32_t>(stoul(RowData.at(3))));
+		Data.Num = static_cast<uint64_t>(stoul(RowData.at(1)));
+		Data.Coord = Coordinate(static_cast<uint64_t>(stoul(RowData.at(2))), static_cast<uint64_t>(stoul(RowData.at(3))));
 		Data.TransferTime = stof(RowData.at(5));
 
 		// todo. convert enum class to string & compare string
@@ -38,8 +38,8 @@ void DataCenter::SetLinkData(const FileDataList& InData)
 		}
 
 		LinkData Data;
-		Data.FromNodeNum = static_cast<uint32_t>(stoul(RowData.at(1)));
-		Data.ToNodeNum = static_cast<uint32_t>(stoul(RowData.at(2)));
+		Data.FromNodeNum = static_cast<uint64_t>(stoul(RowData.at(1)));
+		Data.ToNodeNum = static_cast<uint64_t>(stoul(RowData.at(2)));
 		Data.Length = stof(RowData.at(3));
 		Data.Type = static_cast<LinkType>(stoi(RowData.at(5)));
 		Data.Speed = stof(RowData.at(8));
@@ -58,9 +58,9 @@ void DataCenter::SetRouteData(const FileDataList& InData)
 
 		RouteData Data;
 		Data.Name = RowData.at(0);
-		Data.Order = static_cast<uint32_t>(stoul(RowData.at(1)));
+		Data.Order = static_cast<uint64_t>(stoul(RowData.at(1)));
 		Data.CumDistance = stof(RowData.at(2));
-		Data.Node = static_cast<uint32_t>(stoul(RowData.at(3)));
+		Data.Node = static_cast<uint64_t>(stoul(RowData.at(3)));
 		RouteDataList.emplace_back(Data);
 	}
 }
@@ -76,7 +76,7 @@ void DataCenter::SetOperatingData(const FileDataList& InData)
 
 		OperatingData Data;
 		Data.Name = RowData.at(0);
-		Data.Dispatch = static_cast<uint32_t>(stoul(RowData.at(1)));
+		Data.Dispatch = static_cast<uint64_t>(stoul(RowData.at(1)));
 		Data.Speed = stof(RowData.at(2));
 		OperatingDataList.emplace_back(Data);
 	}
@@ -92,9 +92,9 @@ void DataCenter::SetTrafficVolumeData(const FileDataList& InData)
 		}
 
 		TrafficVolumeData Data;
-		Data.FromNodeNum = static_cast<uint32_t>(stoul(RowData.at(0)));
-		Data.ToNodeNum = static_cast<uint32_t>(stoul(RowData.at(1)));
-		Data.TrafficVolume = static_cast<uint32_t>(stoul(RowData.at(2)));
+		Data.FromNodeNum = static_cast<uint64_t>(stoul(RowData.at(0)));
+		Data.ToNodeNum = static_cast<uint64_t>(stoul(RowData.at(1)));
+		Data.TrafficVolume = static_cast<uint64_t>(stoul(RowData.at(2)));
 		TrafficVolumeDataList.emplace_back(Data);
 	}
 }
@@ -109,8 +109,8 @@ void DataCenter::SetDistanceData(const FileDataList& InData)
 		}
 
 		DistanceData Data;
-		Data.FromNodeNum = static_cast<uint32_t>(stoul(RowData.at(0)));
-		Data.ToNodeNum = static_cast<uint32_t>(stoul(RowData.at(1)));
+		Data.FromNodeNum = static_cast<uint64_t>(stoul(RowData.at(0)));
+		Data.ToNodeNum = static_cast<uint64_t>(stoul(RowData.at(1)));
 		Data.Distance = stof(RowData.at(2));
 		DistanceDataList.emplace_back(Data);
 	}
